Refactor onDropNodes
