+++ src/types/taskNode.ts
@@ -13,6 +13,8 @@ export type NodeAndTaskId = {
   nodeId: string;
 };
 
+export type TaskType = "task" | "input" | "output";
+
 /* Note: Optional callbacks will cause TypeScript to break when applying the callbacks to the Nodes. */
 export interface TaskNodeCallbacks {
   setArguments: (args: Record<string, ArgumentType>) => void;
